Renamed user history to search_history and fixed an issue where search history received the entire user object and not just the username
