+++ seasoned_api/src/webserver/controllers/user/search_history.js
